multibody: Move helper into test_utilities
All testonly cc_library should be clearly marked as either shared
code (via placment in a test_utilities subdirectory), or private (via
private visibility).  Since this helper is shared across packages, it
should be in test_utilities.

Also remove spurious dependency in //multibody/tree, so that no rules
within //multibody/tree depend directly on the plant.
