+++ multibody/benchmarks/free_body/test/multibody_plant_free_body_test.cc
@@ -15,10 +15,10 @@
 #include "drake/math/quaternion.h"
 #include "drake/math/rotation_matrix.h"
 #include "drake/multibody/benchmarks/free_body/free_body.h"
+#include "drake/multibody/test_utilities/floating_body_plant.h"
 #include "drake/multibody/tree/body.h"
 #include "drake/multibody/tree/mobilizer.h"
 #include "drake/multibody/tree/multibody_tree.h"
-#include "drake/multibody/tree/test/floating_body_plant.h"
 #include "drake/systems/analysis/runge_kutta3_integrator.h"
 #include "drake/systems/analysis/simulator.h"
