+++ multibody/test_utilities/floating_body_plant.h
